+++ src/core/interaction.js
